move to common, rename according to function, add additional comment
